UnitTestPkg/Bug51: Add DXE SMM Driver to reproduce bug #51
https://bugzilla.tianocore.org/show_bug.cgi?id=51

This DXE SMM driver uses the SmmPeriodicTimerLib to
create a periodic SMI that is triggered every 8 seconds.
The SMI handler prints a DEBUG() message, so the triggering
of the SMI handler can be seen.  After the periodic SMI
handler is triggered 4 times (32 seconds), the SMI handler
unregisters itself, and the periodic SMI handler should
not be triggered again, and no more DEBUG() messages
from the periodic SMI handle are displayed.

Bug #51 is a use after free issue in the linked list
managed by the QncSmmDispatcher.  This test should
generate unexpected behavior or an an ASSERT() when the
linked list element associated with the SMI handler that
is unregistered is removed from the list in the middle
of processing the list.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Michael Kinney <michael.d.kinney@intel.com>