MdePkg/Include: Fix XCODE5 vaarg issue for X64
The XCODE5 tool chain is not compatible with the
vaarg builtins with _ms_ for X64 builds.

This is a temporary workaround to remove the use
of the _ms_ vaarg builtins from Base.h.

A complete fix will likely use #ifdef statements
to detect the XCODE tool chain and select the
correct vaarg builtins.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>