Vlv2TbltDevicePkg/AcpiPlatform: Change Size from UINT32 to UINTN
Change the type of the local variable Size in the function
AcpiPlatformEntryPoint() from UINT32 to UINTN.  This local
variable is passed into the function ReadSection() that returns
a value of type UINTN.  Using type UINT32 can potentially
corrupt an adjacent local on the stack.  With some GCC builds
the local variable AcpiSupport was being set to zero when
ReadSection() was called.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.1