Vlv2TbltDevicePkg/PlatformFlashAccessLib: Add error return
If the FLASH device cannot be erased or written, then the
flag FlashError is set.

Add the missing logic to return an error if FlashError
is TRUE.  The missing logic was found from a GCC build that
noticed a local variable was declared and assigned, but was
not being used.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.1