OvmfPkg/ResetSystemLib: Change capsule reset to warm reset
Change S3 with immediate wake to a warm reset.  The S3 with
immediate wake does not wake up immediately.  QEMU waits for a
wake event from keyboard or an RTC wakeup timer.

Both cold reset and warm reset for QEMU preserve all system
memory contents, so use a warm reset instead of S3 with an
immediate wake, so a capsule will be processed without waiting
for a keyboard action.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>