FmpDevicePkg/FmpDxe: Fix SetLowestSupportedVersionInVariable()
The change that merged 4 UEFI Variables per device into a single
UEFI variable per device introduced a change in the behavior of
SetLowestSupportedVersionInVariable().  The UEFI Variable should
be updated if the current value is less than the new value.

The BOOLEAN local variable Update in all 4 Setxxx() functions
is also not initialized before use.  Initialize Update to
FALSE in all 4 functions.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>