EmulatorPkg/Sec: Change local variable scope for XCODE5
The local variable PpiArray[10] is declared in middle
of the SEC module _ModuleEntryPoint().  When building
for XCODE5 with optimizations enabled, this results in
corruption and an exception.  The fix is to move the
declaration of PpiArray[10] to the standard location
at the beginning of the function so the storage for
this local variable is allocated for the entire
lifetime of the function.

Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Signed-off-by: Andrew Fish <afish@apple.com>