CryptoPkg/CryptoPkg.dsc: Add build of Crypto libraries/modules
Based on the following package with changes to merge into
CryptoPkg.

https://github.com/microsoft/mu_plus/tree/dev/201908/SharedCryptoPkg

Add Crypto library instances and modules that consume/produce
the EDK II Crypto Protocols/PPIs to the CryptoPkg DSC file.

The default build of CryptoPkg builds all libraries and
modules with all services in the EDK II Crypto Protocols/PPIs
enabled.

The CRYPTO_SERVICES define is added that supports the following
settings.

* ALL             - Build all components.  Protocols and PPIs publish all
                    services.
* NONE            - Protocols/PPIs publish no services.  Used to verify
                    compiler/linker optimizations are working correctly.
* MIN_PEI         - Build PEIM with PPI that publishes minimum required
                    services.
* MIN_DXE_MIN_SMM - Build DXE and SMM drivers with Protocols that publish
                    minimum required services.

Cc: Jian J Wang <jian.j.wang@intel.com>
Cc: Xiaoyu Lu <xiaoyux.lu@intel.com>
Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>