.github/workflow: Add AssignReviewers.py script
Collect GitHub IDs and email address of maintainers
and reviewers for each commit in a pull request.  The
file Maintainer.txt and GetMaintainers.py are used to
collect this information on each commit.  A comment
is added to each commit with the reviewer assignments
and the set of reviewers from all the commits are
assigned to the pull request.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>